Reorganize file structure and naming conventions
