+++ src/components/Nav/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Nav = () => {
-return <h1>NAV</h1>;
+  return <h1>NAV</h1>;
 }
 
 export default Nav;
