+++ src/components/Terminal/index.jsx
@@ -12,14 +12,14 @@ class Terminal extends React.Component {
   }
 
   handleInput = (event) => {
-    this.setState({ command: event.target.value });
+    this.setState({command: event.target.value});
   }
 
   submitCommand = (event) => {
     if (event.key === 'Enter') {
       const splitCommand = this.state.command.split(' ');
       this.props.handleNewCommand(splitCommand);
-      this.setState({ command: '' });
+      this.setState({command: ''});
     }
   }
 
@@ -31,15 +31,15 @@ class Terminal extends React.Component {
 
   render() {
     return (
-      <div className='terminal-window'>
+      <div className="terminal-window">
         <p>Welcome to the Turing Terminal!</p>
-        { this.showPreviousOutput() }
+        {this.showPreviousOutput()}
         <label>~</label>
         <input
-          className='prompt-input'
-          value={ this.state.command }
-          onChange={ this.handleInput }
-          onKeyPress={ this.submitCommand } />
+          className="prompt-input"
+          value={this.state.command}
+          onChange={this.handleInput}
+          onKeyPress={this.submitCommand} />
       </div>
     );
   }
