+++ src/components/Terminal/terminal.css
